Grouped all the single new notification tests into one, refactored how user variables were handled
