--- packages/pn-pa-webapp/cypress/fixtures/recipients/cesare.json
@@ -1,11 +1,0 @@
-{
-  "firstname": "Gaio Giulio",
-  "lastname": "Cesare",
-  "taxId": "CSRGGL44L13H501E",
-  "address": "Via Appia",
-  "houseNumber": "1",
-  "municipality": "Roma",
-  "province": "RM",
-  "zip": "00186",
-  "foreignState": "Italia"
-}
